Add TextRenderer
- This class contains all static rendering methods for Text from the RenderEngine.
