+++ src/de/gurkenlabs/litiengine/graphics/DebugRenderer.java
@@ -96,9 +96,9 @@ public class DebugRenderer {
     g.setFont(g.getFont().deriveFont(Font.PLAIN, 4f));
     final int x = (int) Game.getCamera().getViewPortDimensionCenter(entity).getX() + 10;
     final int y = (int) Game.getCamera().getViewPortDimensionCenter(entity).getY();
-    RenderEngine.drawText(g, Integer.toString(entity.getMapId()), x, y);
+    TextRenderer.drawText(g, Integer.toString(entity.getMapId()), x, y);
     final String locationString = "[x:" + new DecimalFormat("##.##").format(entity.getLocation().getX()) + ";y:" + new DecimalFormat("##.##").format(entity.getLocation().getY()) + "]";
-    RenderEngine.drawText(g, locationString, x, y + 5.0);
+    TextRenderer.drawText(g, locationString, x, y + 5.0);
   }
 
   private static void drawTileBoundingBox(final Graphics2D g, final IMap map, final Point2D location) {
@@ -114,7 +114,7 @@ public class DebugRenderer {
     g.setFont(g.getFont().deriveFont(3f));
     final FontMetrics fm = g.getFontMetrics();
     final Point2D relative = Game.getCamera().getViewPortLocation(playerTile.getX(), playerTile.getY());
-    RenderEngine.drawText(g, locationText, (float) (relative.getX() + playerTile.getWidth() + 3), (float) (relative.getY() + fm.getHeight()));
+    TextRenderer.drawText(g, locationText, (float) (relative.getX() + playerTile.getWidth() + 3), (float) (relative.getY() + fm.getHeight()));
 
     final List<ITile> tiles = MapUtilities.getTilesByPixelLocation(map, location);
     final StringBuilder sb = new StringBuilder();
@@ -122,6 +122,6 @@ public class DebugRenderer {
       sb.append("[gid: " + tile.getGridId() + "] ");
     }
 
-    RenderEngine.drawText(g, sb.toString(), (float) (relative.getX() + playerTile.getWidth() + 3), (float) (relative.getY() + fm.getHeight() * 2 + 2));
+    TextRenderer.drawText(g, sb.toString(), (float) (relative.getX() + playerTile.getWidth() + 3), (float) (relative.getY() + fm.getHeight() * 2 + 2));
   }
 }
